:recycle: (client/EncodeView) Refactor the encoding page
